:recycle: 이메일 인증 API(MailConfig이동, Value어노테이션 사용)
