+++ src/main/java/org/finmate/config/RootConfig.java
@@ -78,22 +78,5 @@ public class RootConfig {
         return manager;
     }
 
-    @Bean
-    public JavaMailSender mailSender() {
-        JavaMailSenderImpl mailSender = new JavaMailSenderImpl();
-        mailSender.setHost("smtp.gmail.com");
-        mailSender.setPort(587);
-        mailSender.setUsername("finmate456@gmail.com");
-        mailSender.setPassword("dnjotleisqzywsmb"); // Gmail 앱 비밀번호
-
-        Properties props = mailSender.getJavaMailProperties();
-        props.put("mail.transport.protocol", "smtp");
-        props.put("mail.smtp.auth", "true");
-        props.put("mail.smtp.starttls.enable", "true");
-        props.put("mail.debug", "true");
-
-        return mailSender;
-    }
-
 
 }
